+++ lib/qunit-dom-globals.ts
@@ -1,0 +1,5 @@
+/* global QUnit */
+
+import install from './install';
+
+install(QUnit);
