+++ lib/qunit-dom-modules.ts
@@ -1,0 +1,4 @@
+import * as QUnit from 'qunit';
+import install from './install';
+
+install(QUnit);
