+++ rollup.config.js
@@ -1,19 +1,19 @@
 import typescript from 'rollup-plugin-typescript2';
 
-export default {
-  input: 'lib/qunit-dom.ts',
+const typescriptConfiguration = {
+  exclude: [
+    'node_modules/**',
+    'lib/**/__tests__/**',
+    'lib/**/*.test.ts',
+    'lib/helpers/test-assertions.ts',
+  ],
+};
+
+const iifeBundle = {
+  input: 'lib/qunit-dom-globals.ts',
 
   external: ['qunit'],
-  plugins: [
-    typescript({
-      exclude: [
-        'node_modules/**',
-        'lib/**/__tests__/**',
-        'lib/**/*.test.ts',
-        'lib/helpers/test-assertions.ts',
-      ],
-    }),
-  ],
+  plugins: [typescript(typescriptConfiguration)],
 
   output: {
     file: 'dist/qunit-dom.js',
@@ -21,3 +21,18 @@ export default {
     sourcemap: true,
   },
 };
+
+const esBundle = {
+  input: 'lib/qunit-dom-modules.ts',
+
+  external: ['qunit'],
+  plugins: [typescript(typescriptConfiguration)],
+
+  output: {
+    file: 'dist/qunit-dom.es.js',
+    format: 'es',
+    sourcemap: true,
+  },
+};
+
+export default [iifeBundle, esBundle];
